clone conflict resolved

+++ set_aws_shortcuts.sh
@@ -9,17 +9,17 @@ fi
 
 printf "\nStarted setting the project..."
 
+if ! [[ -z "$project" ]]; then
+    rm -rf "$HOME/.$project"
+fi
+
 mkdir -p "$HOME/.$project"
 mkdir -p "$HOME/.$project/logs"
+
 cd "$HOME/.$project"
 
-if [[ -d "$project" ]]; then
-    cd "$project"
-    git pull --quiet origin master >/dev/null
-else
-    git clone --quiet https://github.com/sunil-saini/"$project".git >/dev/null
-    cd "$project"
-fi
+git clone --quiet https://github.com/sunil-saini/"$project".git >/dev/null
+cd "$project"
 
 cron="$HOME/.$project/$project/cron.sh"
 
@@ -28,18 +28,21 @@ printf "\nInstalling pip dependencies..."
 python -m pip install --ignore-installed -q -r requirements.txt --user
 
 printf "\nStarted collecting data from AWS, it may take few minutes...\n"
-python driver.py
 
-crontab -l > current_cron
-cron_line="0 */2 * * * /bin/bash $cron"
-if grep -Fxq "$cron_line" current_cron; then
-    echo "Cron already set, skipping"
+if python driver.py ; then
+    crontab -l > current_cron
+    cron_line="0 */2 * * * /bin/bash $cron"
+    if grep -Fxq "$cron_line" current_cron; then
+        echo "Cron already set, skipping"
+    else
+        echo "0 */2 * * * /bin/bash $cron" >> current_cron
+        crontab current_cron
+        rm current_cron
+        echo "Cron set successfully to keep updating data from AWS periodically"
+    fi
+
+    source "$HOME/.$project/.aliases"
+    echo "Project set successfully, Open new terminal tab and enjoy the shortcut commands"
 else
-    echo "0 */2 * * * /bin/bash $cron" >> current_cron
-    crontab current_cron
-    rm current_cron
-    echo "Cron set successfully to keep updating data from AWS periodically"
+    echo "Error(s) in setting the project, please fix above mentioned error and run again"
 fi
-
-source "$HOME/.$project/.aliases"
-echo "Project set successfully, Open new terminal tab and enjoy the shortcut commands"
